Refactor SSD1305 i2c display library and add thread-safety #10.
