+++ lib/Display/Display.h
@@ -1,0 +1,11 @@
+//
+// Display
+//
+
+#ifndef DISPLAY_H
+#define DISPLAY_H
+
+void init_display(void);
+void draw_display_demo_task(void *pvParameter);
+
+#endif // DISPLAY_H
